+++ app/Services/LoadBalancingService.php
@@ -10,12 +10,6 @@ class LoadBalancingService
 {
     private $servers;
 
-    private int $participantWeight = 1;
-
-    private int $audioWeight = 2;
-
-    private int $videoWeight = 3;
-
     public function setServerPool(ServerPool $serverPool)
     {
         $this->servers = $serverPool->servers;
@@ -32,15 +26,9 @@ class LoadBalancingService
             ->where('status', ServerStatus::ENABLED)
             ->where('recover_count', '>=', config('bigbluebutton.server_healthy_threshold'))
             ->where('error_count', '=', 0)
+            ->whereNotNull('load')
             ->sortBy(function (Server $server) {
-                // Experimental
-                // Have video factor 3, audio factor 2 and just listening factor 1
-                $videoLoad = $server->video_count * $this->videoWeight;
-                $voiceLoad = $server->voice_participant_count * $this->audioWeight;
-                $participantLoad = ($server->participant_count - $server->voice_participant_count) * $this->participantWeight;
-                $totalLoad = $videoLoad + $voiceLoad + $participantLoad;
-
-                return $totalLoad / $server->strength;
+                return $server->load / $server->strength;
             })
             ->first();
     }