+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show server connection status in server list ([#911])
 - End detached meetings after server failure ([#911])
 - Added config options for server health `BBB_SERVER_HEALTHY_THRESHOLD` and `BBB_SERVER_UNHEALTHY_THRESHOLD` ([#911])
+- Added config option for server load calculation `BBB_LOAD_MIN_USER_COUNT` and `BBB_LOAD_MIN_USER_INTERVAL` ([#956])
+- Added plugin to customize the server load calculation ([#956])
 
 ### Changed
 - Renamed env option OWN_ROOMS_PAGINATION_PAGE_SIZE to ROOM_PAGINATION_PAGE_SIZE (OWN_ROOMS_PAGINATION_PAGE_SIZE deprecated) ([#373])
@@ -38,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactored frontend from Vue2 to Vue3, replacing BootstrapVue with PrimeVue ([#883])
 - Upgrade to Laravel 11 ([#927])
 - Refactor server health, making it more robust against temporary connection issues ([#911])
+- Calculation of server load uses the participants amount, during starting phase using a configurable min. amount ([#956])
 
 ### Fixed
 - Issue frontend recompiled on every restart due to a hashing issue ([#792])
@@ -77,6 +80,7 @@ You can find the changelog for older versions there [here](https://github.com/TH
 [#905]: https://github.com/THM-Health/PILOS/pull/905
 [#911]: https://github.com/THM-Health/PILOS/pull/911
 [#927]: https://github.com/THM-Health/PILOS/pull/927
+[#956]: https://github.com/THM-Health/PILOS/pull/956
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/v3.0.1...develop
 [v3.0.0]: https://github.com/THM-Health/PILOS/releases/tag/v3.0.0